Refactor api/utils/__init__.py
